services/telegram-bot/requirements.txt
# Telegram Bot Dependencies
aiogram==3.3.0
python-dotenv==1.0.0

# HTTP and API clients
httpx==0.25.2
aiohttp==3.9.1

# Database
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.13.1

# Search and embeddings
sentence-transformers==4.1.0
faiss-cpu==1.7.4
numpy==1.24.3

# Text processing
nltk==3.8.1
pymorphy2==0.9.1

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
loguru==0.7.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1 